# Fix multi-instance benchmarking for fp16